refactoring and adding to folders
